requirements.txt
numpy==1.20.1
Pillow==8.1.2
protobuf==3.14.0
six==1.15.0
PyQt5==5.15.4
matplotlib~=3.3.4